# docs/swagger.yaml
basePath: /
definitions:
  categoriesCreate.Request:
    properties:
      answer:
        type: string
      description:
        type: string
      title:
        type: string
    required:
    - answer
    - description
    - title
    type: object
...
  domain.Category:
    properties:
      answer:
        type: string
      description:
        description: Detailed description of the categories
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  domain.Complaint:
    properties:
      answer:
        example: Complaint resolved
        type: string
      barcode:
        example: 242590
        type: integer
      category:
        $ref: '#/definitions/domain.Category'
      created_at:
        example: "2025-04-21T12:00:00Z"
        type: string
      id:
        example: 71cb6c7d-5c1c-4a5f-bc08-204b6d435c25
        type: string
      message:
        example: Too noisy in the dorm
        type: string
      status:
        example: approved
        type: string
      updated_at:
        example: "2025-04-21T14:00:00Z"
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    required:
    - statusCode
    type: object
...
  updateCategory.Request:
    properties:
      answer:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    required:
    - answer
    - description
    - title
    type: object
host: complaints-api.yeunikey.dev
info:
  contact:
    email: quanaimadi@.gmail.com
    name: API Support
  description: This is a server for managing complaints and categories.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Complaint Server API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      description: 'Создает новую категорию жалоб с необходимыми данными: название,
        описание и ответ.'
      parameters:
      - description: Данные категории
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categoriesCreate.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно создана
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Ошибка валидации или декодирования данных
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет категорию жалоб по переданному идентификатору.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно удалена
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Неверный запрос (не указан ID) или некорректный ID
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Обновляет информацию о категории жалоб. Требуется предоставить
        ID категории и новые данные (название, описание и ответ).
      parameters:
      - description: Данные категории
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updateCategory.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно обновлена
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Ошибка валидации или декодирования данных
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Categories
...
  /categories:
    get:
      consumes:
      - application/json
      description: Возвращает список всех категорий жалоб, доступных в системе.
      produces:
      - application/json
      responses:
        "200":
          description: List of all categories
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
        "500":
          description: Internal server error while fetching categories
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает категорию по уникальному идентификатору (ID).
      parameters:
      - description: Category ID (unique identifier of the categories)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error while fetching the categories
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Categories
  /categories/{id}/complaints:
    get:
      consumes:
      - application/json
      description: Retrieve all complaints that belong to a specific categories based
        on its unique identifier (Category ID).
      parameters:
      - description: Category UUID (unique identifier of the categories)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of complaints associated with the given categories
          schema:
            items:
              $ref: '#/definitions/domain.Complaint'
            type: array
        "400":
          description: Invalid categories ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: No complaints found for the given categories
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error while fetching complaints
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get complaints by categories UUID
      tags:
      - Complaints
...
schemes:
- https
swagger: "2.0"